+++ op-test/components/l1cl/options.go
@@ -1,8 +1,8 @@
 package l1cl
 
 import (
-	test "github.com/ethereum-optimism/optimism/op-test"
 	"github.com/ethereum-optimism/optimism/op-test/components/l1el"
+	"github.com/ethereum-optimism/optimism/op-test/test"
 )
 
 type Settings struct {
