+++ op-test/components/l1cl/types.go
@@ -3,9 +3,9 @@ package l1cl
 import (
 	"github.com/stretchr/testify/require"
 
-	test "github.com/ethereum-optimism/optimism/op-test"
 	"github.com/ethereum-optimism/optimism/op-test/components/l1"
 	"github.com/ethereum-optimism/optimism/op-test/components/l1el"
+	"github.com/ethereum-optimism/optimism/op-test/test"
 )
 
 type L1CL interface {
