+++ op-test/components/l1el/options.go
@@ -1,8 +1,8 @@
 package l1el
 
 import (
-	test "github.com/ethereum-optimism/optimism/op-test"
 	"github.com/ethereum-optimism/optimism/op-test/components/l1"
+	"github.com/ethereum-optimism/optimism/op-test/test"
 )
 
 type Settings struct {
