+++ op-test/components/l2batcher/options.go
@@ -1,8 +1,8 @@
 package l2batcher
 
 import (
-	test "github.com/ethereum-optimism/optimism/op-test"
 	"github.com/ethereum-optimism/optimism/op-test/components/l2cl"
+	"github.com/ethereum-optimism/optimism/op-test/test"
 )
 
 type Settings struct {
