+++ op-test/test/iface.go
@@ -1,4 +1,4 @@
-package op_test
+package test
 
 import (
 	"context"
