+++ op-test/test/planner.go
@@ -1,4 +1,4 @@
-package op_test
+package test
 
 import (
 	"context"
